drf_backend/requirements.txt
asgiref
Django
djangorestframework
djongo
dnspython
pymongo
python-decouple
sqlparse
django
typing_extensions
django-cors-headers